+++ src/commands/archive.ts
@@ -17,7 +17,7 @@ import { hasValidInput } from '../validators';
 // --------------------------------------------------
 // DECLARE FUNCTIONS
 // --------------------------------------------------
-function archive( INPUT: GoalistInput, ARGS: GoalistArgs, config: GoalistConfig ): Promise<GoalistLog> {
+function archive( INPUT: GoalistInput, ARGS: GoalistArgs, config: GoalistConfig ): GoalistLog {
 	let identifier = INPUT[ 0 ] || null;
 
 	// Read in 'source' and 'target' data based on presence of '--active' flag.
@@ -42,13 +42,7 @@ function archive( INPUT: GoalistInput, ARGS: GoalistArgs, config: GoalistConfig 
 	}
 
 	// Update `active` key.
-	if ( isActive ) {
-		config.debugger.log( `Deactivating the following task: ${identifier}` );
-		goal.active = false;
-	} else {
-		config.debugger.log( `Activating the following task: ${identifier}` );
-		goal.active = true;
-	}
+	goal.active = isActive ? false : true;
 
 	// Remove target goal; update 'source' and 'target' data.
 	delete goals[ identifier ];