+++ src/commands/backup.ts
@@ -7,7 +7,8 @@ import * as fs from 'fs';
 // Vendors
 
 // Project
-import { GoalistLog, GoalistArgs, GoalistConfig, GoalistInput } from '../interfaces';
+import makeCommand from './utils';
+import { CommandPayload, GoalistLog, GoalistArgs, GoalistConfig, GoalistInput } from '../interfaces';
 
 // --------------------------------------------------
 // DECLARE VARS
@@ -16,20 +17,20 @@ import { GoalistLog, GoalistArgs, GoalistConfig, GoalistInput } from '../interfa
 // --------------------------------------------------
 // DECLARE FUNCTIONS
 // --------------------------------------------------
-export default function backup( INPUT: GoalistInput, ARGS: GoalistArgs, config: GoalistConfig ): Promise<GoalistLog> {
-	return new Promise( ( resolve, reject ) => {
-		let log = ARGS.archive ? config.utils.getLog( 'archive' ) : config.utils.getLog( 'active' );
+function backup( INPUT: GoalistInput, ARGS: GoalistArgs, config: GoalistConfig ): CommandPayload<GoalistLog> {
+	let log = ARGS.archive ? config.utils.getLog( 'archive' ) : config.utils.getLog( 'active' );
 
-		try {
-			fs.writeFileSync( `${config.utils.getBakPath()}/goalist_${ARGS.archive ? 'archive' : 'active'}_${new Date().getTime()}.bak`, JSON.stringify( log ), 'utf8' );
-			config.debugger.log( `Successfully backed up log: ${ARGS.archive ? 'archive' : 'active'}` );
-			resolve( log );
-			return;
-		} catch ( err ) {
-			config.debugger.log( `Failed to back up log: ${ARGS.archive ? 'archive' : 'active'}` );
-			reject( err );
-			return;
-		}
+	try {
+		fs.writeFileSync( `${config.utils.getBakPath()}/goalist_${ARGS.archive ? 'archive' : 'active'}_${new Date().getTime()}.bak`, JSON.stringify( log ), 'utf8' );
 
-	} );
+		return {
+			msg: `Successfully backed up log: ${ARGS.archive ? 'archive' : 'active'}`,
+			payload: log,
+		};
+	} catch ( err ) {
+		throw new Error( `Failed to back up log: ${ARGS.archive ? 'archive' : 'active'}` );
+	}
 }
+
+export default makeCommand( backup, [ () => Promise.resolve() ] );
+