+++ src/commands/remove.ts
@@ -8,7 +8,7 @@ import * as fs from 'fs';
 
 // Project
 import makeCommand from './utils';
-import { GoalistLog, GoalistArgs, GoalistConfig, GoalistInput } from '../interfaces';
+import { Goal, GoalistLog, GoalistArgs, GoalistConfig, GoalistInput } from '../interfaces';
 import { hasValidInput } from '../validators';
 
 // --------------------------------------------------
@@ -18,7 +18,7 @@ import { hasValidInput } from '../validators';
 // --------------------------------------------------
 // DECLARE FUNCTIONS
 // --------------------------------------------------
-function remove( INPUT: GoalistInput, ARGS: GoalistArgs, config: GoalistConfig ): Promise<GoalistLog> {
+function remove( INPUT: GoalistInput, ARGS: GoalistArgs, config: GoalistConfig ): Goal {
 	let identifier = INPUT[ 0 ] || null;
 
 	let log = ARGS.archive ? config.utils.getLog( 'archive' ) : config.utils.getLog( 'active' );
@@ -27,18 +27,13 @@ function remove( INPUT: GoalistInput, ARGS: GoalistArgs, config: GoalistConfig )
 	let userConf = null;
 
 	if ( !goal ) {
-		let err = `Whoops, failed to find a goal which matches the following identifier: ${identifier}`;
-
-		config.debugger.log( err );
-		throw new Error( err );
+		throw new Error( `Whoops, failed to find a goal which matches the following identifier: ${identifier}` );
 	}
 
 	// Proceed if:
 	// - we're not running in CLI mode;
 	// - or the `--force` flag was provided.
 	if ( !config.cli || ARGS.force ) {
-		config.debugger.log( `Removing task: ${identifier}` );
-
 		for ( let key in goals ) {
 			if ( goals[ key ] === goal ) {
 				delete goals[ key ];
@@ -48,10 +43,9 @@ function remove( INPUT: GoalistInput, ARGS: GoalistArgs, config: GoalistConfig )
 		// Write new data back to file system.
 		config.utils.writeLog( ARGS.archive ? 'archive' : 'active', JSON.stringify( log ) );
 
-		return log;
+		return goal;
 	} else {
-		config.debugger.log( 'This is a destructive action and can only be executed if the `--force` flag is provided. Aborting.' );
-		return;
+		throw new Error( 'This is a destructive action and can only be executed if the `--force` flag is provided. Aborting.' );
 	}
 }