+++ src/loggers/add.ts
@@ -1,0 +1,5 @@
+import { Goal } from '../interfaces';
+
+export default function add(goal: Goal, goalist) {
+	goalist.config.debugger.log( `Successfully created goal: ${goal.id}` );
+}