+++ test/commands/list.test.js
@@ -50,94 +50,91 @@ test( 'it should return a Promise', ( t ) => {
   t.is( result instanceof Promise, true );
 } );
 
-test( 'resolve with the log', ( t ) => {
-  const log = { goals: {} };
+test( 'resolve with an array of goal objects', ( t ) => {
+  const goal = { id: 1 };
+  const log = { goals: { '1': goal } };
   const getLog = config.utils.getLog = sinon.spy( () => log );
 
   return list( [], {}, config ).then( ( data ) => {
-    t.is( data, log );
+    t.deepEqual( data, [ goal ] );
   } );
 } );
 
 test( 'it should extract the "archive" log', ( t ) => {
   const getLog = config.utils.getLog = sinon.spy( () => ( { goals: {} } ) );
 
-  list( [], { archive: true }, config );
-
-  t.is( getLog.calledWith( 'archive' ), true);
+  return list( [], { archive: true }, config ).
+    then( () => {
+      t.is( getLog.calledWith( 'archive' ), true);
+    } );
 } );
 
 test( 'it should extract the "active" log', ( t ) => {
   const getLog = config.utils.getLog = sinon.spy( () => ( { goals: {} } ) );
 
-  list( [], {}, config );
-
-  t.is( getLog.calledWith( 'active' ), true );
+  return list( [], {}, config )
+    .then( () => {
+      t.is( getLog.calledWith( 'active' ), true );
+    } );
 } );
 
-test( 'should display the "id" and "title" data for each goal', ( t ) => {
+test( 'should return goals with "id" and "title" properties by default', ( t ) => {
   const goal = { id, title };
-  const getLog = config.utils.getLog = sinon.spy( () => ( { goals: { [id]: goal } } ) );
-  const log = config.debugger.log = sinon.spy();
+  const getLog = config.utils.getLog = sinon.spy( () => ( { goals: { [ id ]: goal } } ) );
 
-  list( [], {}, config );
-
-  t.is( log.args[0][0].includes( id ), true );
-  t.is( log.args[1][0].includes( title ), true );
+  return list( [], {}, config )
+    .then( ( [ goal ] ) => {
+      t.true( 'id' in goal );
+      t.true( 'title' in goal );
+    } );
 } );
 
-test( 'should display the supplementary data for each goal when the "show" argument is provided', ( t ) => {
+test( 'should return goals with supplementary properties when the "show" argument is provided', ( t ) => {
   const goal = { id, title, description };
-  const getLog = config.utils.getLog = sinon.spy( () => ( { goals: { [id]: goal } } ) );
-  const log = config.debugger.log = sinon.spy();
+  const getLog = config.utils.getLog = sinon.spy( () => ( { goals: { [ id ]: goal } } ) );
 
-  list( [], { show: 'description' }, config );
-
-  t.is( log.args[0][0].includes( id ), true );
-  t.is( log.args[1][0].includes( title ), true );
-  t.is( log.args[2][0].includes( description ), true );
+  return list( [], { show: 'description' }, config )
+    .then( ( [ goal ] ) => {
+      t.true( 'id' in goal );
+      t.true( 'title' in goal );
+      t.true( 'description' in goal );
+    } );
 } );
 
 test( 'should display all of the data for each goal when the "all" argument is provided', ( t ) => {
   const goal = { id, title, description, complete, active };
   const getLog = config.utils.getLog = sinon.spy( () => ( { goals: { [ id ]: goal } } ) );
-  const log = config.debugger.log = sinon.spy();
 
-  list( [], { all: true }, config );
-
-  t.is( log.args[0][0].includes( id ), true );
-  t.is( log.args[1][0].includes( title ), true );
-  t.is( log.args[2][0].includes( description ), true );
-  t.is( log.args[3][0].includes( complete ), true );
-  t.is( log.args[4][0].includes( active ), true );
+  return list( [], { all: true }, config )
+    .then( ( [ goal ] ) => {
+      t.true( 'id' in goal );
+      t.true( 'title' in goal );
+      t.true( 'description' in goal );
+      t.true( 'complete' in goal );
+      t.true( 'active' in goal );
+    } );
 } );
 
 test( 'it should allow goals to be filtered by category', ( t ) => {
   const goal1 = { title, category };
   const goal2 = { title: title2, category: category2 };
   const getLog = config.utils.getLog = sinon.spy( () => ( { goals: { '1': goal1, '2': goal2 } } ) );
-  const log = config.debugger.log = sinon.spy();
 
-  list( [], { category }, config );
-
-  // Flatten `log()` arugments into a 1-dimensional array.
-  const args = log.args.reduce( ( acc, args ) => [...acc, ...args], [] );
-  t.is(args.some( ( arg ) => arg.includes( title ) ), true);
-  t.is(args.every( ( arg ) => !arg.includes( title2 ) ), true);
+  return list( [], { category }, config )
+    .then( ( [ goal ] ) => {
+      t.is( goal, goal1 );
+    } );
 } );
 
 test( 'it should allow goals to be filtered by tag', ( t ) => {
   const goal1 = { title, tags: [ 'foo', 'bar' ] };
   const goal2 = { title: title2, tags: [ 'baz', 'quux' ] };
   const getLog = config.utils.getLog = sinon.spy( () => ( { goals: { '1': goal1, '2': goal2 } } ) );
-  const log = config.debugger.log = sinon.spy();
 
-  list( [], { tags: 'foo,bar' }, config );
-
-  // Flatten `log()` arugments into a 1-dimensional array.
-  const args = log.args.reduce( ( acc, args ) => [...acc, ...args], [] );
-  t.is(args.some( ( arg ) => arg.includes( title ) ), true);
-  t.is(args.every( ( arg ) => !arg.includes( title2 ) ), true);
+  return list( [], { tags: 'foo,bar' }, config )
+    .then( ( [ goal ] ) => {
+      t.is( goal, goal1 );
+    } );
 } );
 
 test( 'it should allow goals to be filtered by category or tag', ( t ) => {
@@ -146,10 +143,8 @@ test( 'it should allow goals to be filtered by category or tag', ( t ) => {
   const getLog = config.utils.getLog = sinon.spy( () => ( { goals: { '1': goal1, '2': goal2 } } ) );
   const log = config.debugger.log = sinon.spy();
 
-  list( [], { tags: 'foo,bar', category: 'baz' }, config );
-
-  // Flatten `log()` arugments into a 1-dimensional array.
-  const args = log.args.reduce( ( acc, args ) => [...acc, ...args], [] );
-  t.is(args.some( ( arg ) => arg.includes( title ) ), true);
-  t.is(args.some( ( arg ) => arg.includes( title2 ) ), true);
+  return list( [], { tags: 'foo,bar', category: 'baz' }, config )
+    .then( ( goals ) => {
+      t.deepEqual( goals, [ goal1, goal2 ] );
+    } );
 } );